+++ spec/haml_lint/linter/unnecessary_string_output_spec.rb
@@ -60,43 +60,49 @@ describe HamlLint::Linter::UnnecessaryStringOutput do
     it { should_not report_lint }
   end
 
-  context 'when script outupts literal string starting with an HTML comment character' do
+  context 'when script outputs literal string starting with an HTML comment character' do
     let(:haml) { '= "/ Something"' }
 
     it { should_not report_lint }
   end
 
-  context 'when script outupts literal string starting with a hash sign' do
+  context 'when script outputs literal string starting with a hash sign' do
     let(:haml) { '= "# Something"' }
 
     it { should_not report_lint }
   end
 
-  context 'when script outupts literal string starting with a dash' do
+  context 'when script outputs literal string starting with a dash' do
     let(:haml) { '= "- Something"' }
 
     it { should_not report_lint }
   end
 
-  context 'when script outupts literal string starting with a equals sign' do
+  context 'when script outputs literal string starting with a equals sign' do
     let(:haml) { '= "= Something"' }
 
     it { should_not report_lint }
   end
 
-  context 'when script outupts literal string starting with a percent sign' do
+  context 'when script outputs literal string starting with a percent sign' do
     let(:haml) { '= "% Something"' }
 
     it { should_not report_lint }
   end
 
-  context 'when script outupts literal string starting with a tilde' do
+  context 'when script outputs literal string starting with a tilde' do
     let(:haml) { '= "~ Something"' }
 
     it { should_not report_lint }
   end
 
-  context 'when script outupts literal string starting with interpolation' do
+  context 'when script outputs string starting with a special character then interpolation' do
+    let(:haml) { '= "/ #{interpolation}"' }
+
+    it { should_not report_lint }
+  end
+
+  context 'when script outputs literal string starting with interpolation' do
     let(:haml) { '= "#{variable}"' }
 
     it { should report_lint }